refactor: move helpers to utils
